src/fut_with_diag: Reduce overhead when logging disabled
`DiagnoeFuture::poll` either just polls the underlying future when
logging is disabled, or records the start time, polls the future,
records the end time, logs the information and then returns the poll
outcome.

In the former case where logging is disabled, there is no reason to
record start and end time, given that `log_poll` becomes a no-op.